Extract compressor classes
Load compressor implementations explicitly.

Compressor classes will be loaded when `Fluent::Registry` is
initialized. But it is not good practice to depend on the order of
loading files.
